refactor(subagents): split oversized inline tests
Move the command tests to their sibling module after the new regression case crossed the repository's inline-test size boundary.
